+++ src/decorators/index.ts
@@ -4,7 +4,7 @@ export { Authorized } from "./Authorized";
 export { Ctx } from "./Ctx";
 export { Field } from "./Field";
 export { FieldResolver } from "./FieldResolver";
-export { GraphQLArgsType, GraphQLArgumentType } from "./GraphQLArgsType";
+export { GraphQLArgsType } from "./GraphQLArgsType";
 export { GraphQLInputType } from "./GraphQLInputType";
 export { GraphQLInterfaceType } from "./GraphQLInterfaceType";
 export { GraphQLObjectType } from "./GraphQLObjectType";